backend/contracts/deploy.py
"""
EcoSync Smart Contract Deployment Script
Uses web3.py to deploy the EcoSyncMarketplace contract
"""
import json
import os
from web3 import Web3
from eth_account import Account

# Contract bytecode and ABI would be generated from compilation
# This is a placeholder - in production, compile with solc

CONTRACT_ABI = []  # Would be populated from compiled contract
CONTRACT_BYTECODE = ""  # Would be populated from compiled contract


def deploy_contract(
    rpc_url: str = "http://localhost:8545",
    private_key: str = None,
    initial_supply: int = 1000000
):
    """
    Deploy the EcoSync Marketplace contract
    
    Args:
        rpc_url: Ethereum node RPC URL
        private_key: Deployer private key
        initial_supply: Initial EcoToken supply
    
    Returns:
        Contract address
    """
    # Connect to Ethereum node
    w3 = Web3(Web3.HTTPProvider(rpc_url))
    
    if not w3.is_connected():
        raise ConnectionError("Failed to connect to Ethereum node")
    
    print(f"Connected to Ethereum node at {rpc_url}")
    print(f"Chain ID: {w3.eth.chain_id}")
    
    # Create account from private key
    if private_key:
        account = Account.from_key(private_key)
    else:
        # For testing - generate a new account
        account = Account.create()
        print(f"Generated test account: {account.address}")
    
    # In a real deployment, you would:
    # 1. Compile the Solidity contract
    # 2. Get the bytecode and ABI
    # 3. Deploy using web3.py
    
    # For hackathon/demo purposes, we'll create a mock deployment
    print("\n" + "="*60)
    print("EcoSync Smart Contract Deployment")
    print("="*60)
    print("\nNote: This is a demonstration deployment.")
    print("For production, compile the contract with solc and deploy to mainnet/testnet.")
    print("\nContract Details:")
    print(f"  Name: EcoToken (ECO)")
    print(f"  Initial Supply: {initial_supply:,} ECO")
    print(f"  Platform Fee: 1%")
    print(f"  Min Trade: 1 kWh")
    print(f"  Max Trade: 10,000 kWh")
    print("\nFeatures:")
    print("  - P2P Energy Trading")
    print("  - Automated Settlement")
    print("  - ZKP Verification (placeholder)")
    print("  - Oracle-based Validation")
    print("="*60)
    
    # Return mock address
    return "0x" + "0" * 40


def compile_contract():
    """
    Compile the Solidity contract using py-solc-x
    """
    try:
        from solcx import compile_source, install_solc
        
        # Install Solidity compiler
        install_solc('0.8.19')
        
        # Read contract source
        with open('EcoSyncMarketplace.sol', 'r') as f:
            source = f.read()
        
        # Compile
        compiled = compile_source(
            source,
            output_values=['abi', 'bin'],
            solc_version='0.8.19'
        )
        
        # Get contract data
        contract_id, contract_interface = compiled.popitem()
        
        return {
            'abi': contract_interface['abi'],
            'bytecode': contract_interface['bin']
        }
    
    except ImportError:
        print("solcx not installed. Install with: pip install py-solc-x")
        return None
    except Exception as e:
        print(f"Compilation error: {e}")
        return None


if __name__ == "__main__":
    deploy_contract()
