backend/requirements.txt
# EcoSync Backend Dependencies
# Core
numpy>=1.26.0
python-dotenv>=1.0.0

# IoT & MQTT
paho-mqtt>=1.6.1

# FastAPI & Web
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6

# CORS
python-jose[cryptography]>=3.3.0

# AI & LangGraph (optional - for advanced features)
langgraph>=0.0.50
langchain>=0.1.0

# Blockchain (optional - for smart contract interaction)
web3>=6.0.0
eth-account>=0.10.0

# Utilities
aiohttp>=3.9.0
requests>=2.31.0
